[seil] favorite

+++ app/src/main/java/com/example/lllov/projectkjh/BaseActivity.java
@@ -19,7 +19,7 @@ public class BaseActivity extends AppCompatActivity {
     SimpleDateFormat formatDay = new SimpleDateFormat("dd", Locale.KOREA);
 
     public final static HashMap<Integer, String> PLACE_TYPE = new HashMap<>();
-    public static int sUserId;
+    public static long sUserId;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
